LPD-75099 Move useSideNavigationFilter hook to its own file
