+++ modules/apps/application-list/application-list-taglib/src/main/resources/META-INF/resources/js/SideNavigationItem.ts
@@ -1,0 +1,12 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+export interface SideNavigationItem {
+	href?: string;
+	id: string;
+	items?: Array<SideNavigationItem>;
+	label: string;
+	leadingIcon?: string;
+}
