+++ modules/apps/application-list/application-list-taglib/test/js/SideNavigationFilter.test.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {filterItemsByQuery} from '../../src/main/resources/META-INF/resources/js/SideNavigation';
+import {filterItemsByQuery} from '../../src/main/resources/META-INF/resources/js/useSideNavigationFilter';
 
 describe('Single layer items filtering', () => {
 	const singleLayerItems = [
